fix(cmake/cross-build): misc fix 1: fix cmake cross-ios failed caused by df118a87    build static lib for APPLE define for XCODE third_party framework including require
2: megbrain_test/megdnn_test build when MGE_INFERENCE_ONLY=ON
   now u can build megbrain_test/megdnn_test by:
   EXTRA_CMAKE_ARGS="-DMGE_WITH_TEST=ON" ./scripts/cmake-build/xxx.sh
   example macos-cross-ios build megdnn_test for IOS by
   EXTRA_CMAKE_ARGS="-DMGE_WITH_TEST=ON" ./scripts/cmake-build/cross_build_ios_arm_inference.sh

3: reuse host flatc build when cross build mode
GitOrigin-RevId: 132f4bf893567bdb1d54de506449950513a5841f

+++ scripts/cmake-build/utils/utils.sh
@@ -12,18 +12,30 @@ fi
 function build_flatc() {
     BUILD_DIR=$1/build_dir/host_flatc/build
     INSTALL_DIR=$BUILD_DIR/../install
-    if [ -e $BUILD_DIR ];then
-        echo "clean old dir: $BUILD_DIR"
-        rm -rf $BUILD_DIR
-    fi
-    if [ -e $INSTALL_DIR ];then
-        echo "clean old dir: $INSTALL_DIR"
-        rm -rf $INSTALL_DIR
+    REMOVE_OLD_BUILD=$2
+
+    if [ $REMOVE_OLD_BUILD = "true" ]; then
+        if [ -e $BUILD_DIR ];then
+            echo "clean old dir: $BUILD_DIR"
+            rm -rf $BUILD_DIR
+        fi
+        if [ -e $INSTALL_DIR ];then
+            echo "clean old dir: $INSTALL_DIR"
+            rm -rf $INSTALL_DIR
+        fi
+    else
+        echo "strip remove old build"
     fi
 
-    echo "create build dir"
-    mkdir -p $BUILD_DIR
-    mkdir -p $INSTALL_DIR
+    if [ ! -e $BUILD_DIR ];then
+        echo "create build: $BUILD_DIR"
+        mkdir -p $BUILD_DIR
+    fi
+    if [ ! -e $INSTALL_DIR ];then
+        echo "create install $INSTALL_DIR"
+        mkdir -p $INSTALL_DIR
+    fi
+
     cd $BUILD_DIR
     cmake -G "$MAKEFILE_TYPE Makefiles" \
         -DCMAKE_BUILD_TYPE=Release \
